# api/openapi.yaml
openapi: 3.1.0
info:
  title: JobReady API
  version: 1.0.0
  description: |
    JobReady is an AI-powered job application tracker that helps candidates manage their job search end-to-end.

    Core capabilities:
    - Auth & Profiles — JWT-based registration/login; structured candidate profiles (experience, education, skills, certifications, languages).
    - Application Tracking — create applications by pasting a job description; manage status transitions (Applied → Interviewing → Offer → Rejected); auto-link incoming emails to the correct application.
    - Gmail Integration — connect/disconnect Gmail accounts; fetch and store application-related emails; surface follow-up recommendations for stale applications.
    - AI Features — tailored cover letter generation, candidate-role fit analysis with gap identification, and email-based insights — all driven by the candidate profile and target job description.
    - Document Management — store and retrieve AI-generated cover letters and fit analyses.
  contact:
    name: JobReady Support

servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
  - url: https://api.jobready.com
    description: Production

security:
  - BearerAuth: []

tags:
  - name: Auth
    description: Authentication and session management
  - name: Email
    description: Gmail integration — connect/disconnect accounts and read fetched emails

paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: |
        Creates a new account and signs the user in immediately — no separate login step.
        Access and refresh tokens are delivered as HttpOnly cookies (never in the body);
        the response body carries the authenticated user's profile.
      security: []
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created; session cookies set
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionCookies'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error (e.g. password too short)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: |
        Verifies credentials and sets HttpOnly access + refresh cookies.
        Tokens are never returned in the body; the body carries the user's profile.
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful; session cookies set
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionCookies'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Reads the refresh token from the `jr_refresh` HttpOnly cookie, rotates it,
        and sets a fresh access + refresh cookie pair. The old refresh token is
        immediately invalidated. No request body — the cookie carries the token.
      security: []
      operationId: refreshToken
      responses:
        '204':
          description: New session cookies set
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionCookies'
        '401':
          description: Refresh cookie missing, invalid, expired, or already rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and invalidate the session
      description: |
        Revokes the refresh token read from the `jr_refresh` cookie and clears both
        session cookies (`jr_access`, `jr_refresh`). No request body.
      operationId: logout
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        '204':
          description: Logged out; session cookies cleared
          headers:
            Set-Cookie:
              $ref: '#/components/headers/ClearedCookies'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current user info
      description: |
        Returns the authenticated user's profile, resolved from the JWT `sub` claim.
        The access token is taken from the `jr_access` cookie (or, behind the gateway,
        the translated `Authorization: Bearer` header).
      operationId: getMe
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Missing or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/email/connections/gmail/authorize:
    post:
      tags: [Email]
      summary: Start the Gmail OAuth2 consent flow
      description: |
        Returns a Google consent URL the client should redirect the user to. The URL carries a
        signed, short-lived `state` token bound to the authenticated user (`sub` claim) and a
        single-use nonce; the callback verifies it to prevent CSRF/replay and account-binding attacks.
      operationId: authorizeGmail
      responses:
        '200':
          description: Consent URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GmailAuthorizeResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/email/connections/gmail/callback:
    get:
      tags: [Email]
      summary: Gmail OAuth2 redirect callback
      description: |
        Google redirects here with `code` and `state`. The service verifies the `state` signature
        and nonce (rejecting forged/expired/replayed values), derives the user from the verified
        payload, exchanges the code for tokens, stores the connection, and redirects the browser
        back to the frontend. This endpoint is reached by the browser, not called service-to-service,
        so it is unauthenticated — the signed `state` is the trust anchor.
      operationId: gmailOAuthCallback
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Connection stored; browser redirected back to the frontend
          headers:
            Location:
              description: Frontend URL the user is returned to
              schema:
                type: string
        '400':
          description: Missing/invalid code or state, or token exchange failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/email/connections:
    get:
      tags: [Email]
      summary: Get the current user's email connection status
      operationId: getEmailConnection
      responses:
        '200':
          description: Connection status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailConnectionStatus'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Email]
      summary: Disconnect the current user's email account
      description: Deletes the stored OAuth connection for the authenticated user.
      operationId: deleteEmailConnection
      responses:
        '204':
          description: Connection deleted (idempotent — 204 even if none existed)
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/email/messages:
    get:
      tags: [Email]
      summary: List stored emails for the current user
      description: Returns the emails fetched by the background poller for the authenticated user, newest first.
      operationId: listEmailMessages
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Stored emails
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailMessageList'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: jr_access
      description: |
        HttpOnly access-token cookie set by the auth service on login/register.
        Used directly by the auth service; the gateway translates it into an
        `Authorization: Bearer` header for downstream services.

  headers:
    SessionCookies:
      description: |
        Two HttpOnly cookies are set:
        `jr_access` (the access JWT, Path=/, Max-Age, HttpOnly; Secure; SameSite=Strict)
        and `jr_refresh` (the rotating refresh token, Path=/api/v1/auth,
        Max-Age, HttpOnly; Secure; SameSite=Strict).
      schema:
        type: string
      example: jr_access=<jwt>; Path=/; Max-Age=900; HttpOnly; Secure; SameSite=Strict
    ClearedCookies:
      description: Both session cookies are expired (Max-Age=0).
      schema:
        type: string
      example: jr_access=; Path=/; Max-Age=0; HttpOnly; Secure; SameSite=Strict

  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: jane@example.com
        password:
          type: string
          format: password
          minLength: 8

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: jane@example.com
        password:
          type: string
          format: password
          minLength: 8

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Stable user identifier — matches JWT `sub` claim
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_CREDENTIALS
        message:
          type: string
          example: Email or password is incorrect
        details:
          type: object
          additionalProperties: true
          description: Optional field-level validation errors or additional context

    GmailAuthorizeResponse:
      type: object
      required: [authorization_url]
      properties:
        authorization_url:
          type: string
          format: uri
          description: Google consent URL the client should redirect the user to
          example: https://accounts.google.com/o/oauth2/v2/auth?client_id=...&state=...

    EmailConnectionStatus:
      type: object
      required: [connected]
      properties:
        connected:
          type: boolean
          description: Whether the user has a stored email connection
        provider:
          type: string
          enum: [gmail]
          description: Present only when connected
        email_address:
          type: string
          format: email
          description: The connected mailbox address — present only when connected
        connected_at:
          type: string
          format: date-time
          description: When the connection was first established — present only when connected

    EmailMessage:
      type: object
      required: [message_id]
      properties:
        message_id:
          type: string
          description: Provider-side unique message identifier
        subject:
          type: [string, 'null']
        sender:
          type: [string, 'null']
        snippet:
          type: [string, 'null']
        received_at:
          type: [string, 'null']
          format: date-time

    EmailMessageList:
      type: object
      required: [items, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmailMessage'
        limit:
          type: integer
        offset:
          type: integer